experimental/prompt-rewrite-eval/requirements.txt
requests
GitPython
streamlit
jsonlines